feat: add health check and product typing
