+++ frontend/src/stores/product.ts
@@ -1,15 +1,5 @@
 import { defineStore } from 'pinia';
-
-export interface Product {
-  id: number;
-  name: string;
-  slug: string;
-  description?: string;
-  price_cents: number;
-  currency: string;
-  stock: number;
-  image_url?: string;
-}
+import type { Product } from 'src/types/product';
 
 interface Filters {
   search: string;
