+++ frontend/src/types/product.ts
@@ -1,0 +1,10 @@
+export interface Product {
+  id: number;
+  name: string;
+  slug: string;
+  description?: string;
+  price_cents: number;
+  currency: string;
+  stock: number;
+  image_url?: string;
+}
